split data operations and helper functions to data_utils and utils and included check for existence of files
