+++ tools/code/Floods_Fathom3/common.py
@@ -1,5 +1,4 @@
 # This file includes common configuration elements used by the script.
-
 import os
 
 # Load env vars from dotenv file
